requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
scikit-learn==1.3.2
opencv-python==4.8.1.78
reportlab==4.0.7
google-generativeai==0.3.2
numpy==1.24.3
matplotlib>=3.10.0

